fold Extract into TemplateValuesFactory
